fix(tt-482): address pr comments
- move utility functions to utils
- change case_id param to submitter_id and use it
- correctly check node state instead of doc state submitted
- use client_toggled fixture with config flags
