+++ sheepdog/transactions/upload/sub_entities.py
@@ -3,11 +3,11 @@ Subclasses for UploadEntity that handle different types of
 uploaded entites.
 """
 import uuid
-import flask
 
+import psqlgraph
 from indexclient.client import Document
-from sheepdog import utils
 
+from sheepdog import utils
 from sheepdog.transactions.entity_base import EntityErrors
 from sheepdog.transactions.upload.entity import (
     UploadEntity,
@@ -466,6 +466,48 @@ class FileUploadEntity(UploadEntity):
                     is_valid = False
 
         return is_valid
+
+    def get_node_recreate(self):
+        """
+        Create a new node in the old node's place if it exists in indexd. with
+        graph node state = 'released'.
+        """
+
+        nodes = lookup_node(
+            self.transaction.db_driver,
+            self.entity_type,
+            self.entity_id,
+            self.secondary_keys,
+        )
+
+        self.old_uuid = nodes.one().node_id
+        self.file_by_uuid = self.get_file_from_index_by_uuid(self.entity_id)
+        with self.transaction.db_driver.session_scope() as session:
+            for node in nodes:
+                session.delete(node)
+
+        # new node id
+        self.entity_id = str(uuid.uuid4())
+
+        # Fill in default system property values
+        for key, val in self.get_system_property_defaults().iteritems():
+            if self.doc.get(key, None) is None:
+                self.doc[key] = val
+
+        # Create the node and populate its properties
+        cls = psqlgraph.Node.get_subclass(self.entity_type)
+        self.logger.debug('Recreating new {}'.format(cls.__name__))
+        node = cls(self.entity_id)
+
+        # check if open_acl is requested and the node type can be set open
+        if self.doc.get('open_acl', None) and self._config.get('IS_GDC', False):
+            if self.entity_type in POSSIBLE_OPEN_FILE_NODES:
+                node.acl = [u'open']
+            else:
+                node.acl = self.transaction.get_phsids()
+
+        self.action = 'version'
+        return node
 
     def get_file_from_index_by_hash(self):
         """
